requirements.txt
asgiref==3.8.1
Django==4.2.11
django-filter==24.1
djangorestframework==3.15.1
importlib_metadata==7.1.0
Markdown==3.6
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
zipp==3.18.1
